Made README and python_venv more unix friendly
The README and python_venv files were oriented towards Windows users,
but this isn't standard for most Git projects. The commands and
instructions are designed to be used with Git commands and run on a unix
shell ie bash or zsh. Also, the <br> tag was removed in the README file
in favor of double spaces. I also I adjusted the formating to make it
more friendly to read, including lists.